# Speed-up multidim sum