@SsnL 2018-06-28
@pytorchbot retest this please